cloud_function/requirements.txt
requests
Pillow
flask<3.0,>=1.0
tensorflow
numpy
google-cloud-storage